Add some tests for FolderHandler.
Also introduced a neato MockFile class for testing. (There's BuiltinFile
which seems to be the same idea and used by the natlab.refactoring tests,
but it seems kind of complicated...)

It would also be nice to rename lookupSpecializedAll and
getAllSpecialized to be less confusing.

Migrated from Sable/mclab@83c3e19795105f9b97694aaffe74bf5e0d635458