Introduce RefactoringContext.
Refactorings should be aware not just of the AST they're operating on,
but also some extra context (for now the MatlabProgram they're operating
on, and the Project it belongs to).

This is a fairly big change, and some of it is iffy just to make things
build. In particular,

* ExtractFunction needs to be looked at, in particular
  checkPreconditions()
* One RenameVariableTest regarding global variables was commented out,
  and will be fixed soon

Migrated from Sable/mclab@44cdbcfa67d3ad886a605b7c4357d18997bd1f3d