Use Map instead of HashMapFlowMap in ReachingDefs
The union / merge functionality of AbstractFlowMap was added
as a few utility methods in a new MergeUtil class.

(n.b. need to look at the contract for merge() in analyses, because
it's very error prone -- the input and output sets can be the same,
so you need to consider a bunch of cases (both inputs and output
same, all different, one input is same as output, etc.). It would
probably be better if you just returned a new set instead of using
an output parameter, but maybe this was done to have fewer memory
allocations?)

Migrated from Sable/mclab@430c64ca98a7673e10cec9ac9fc55f337321d165