Don't ignore the Natlab translation position map.
This should be make the line numbers more accurate in the
presence of e.g. missing ends and nested functions.

It doesn't seem to be perfect, though. But at least the line
numbers that are wrong seem to be closer.

Migrated from Sable/mclab@973146d47a341e2785c49d5be8bbdccaaab2d16d