Tentative fix for "unexpected HELP_COMMENT" issue.
Background: in the Natlab parser, comments starting with two percent
signs are specially recognized "help comments", because e.g. Matlab's
"help" function recognizes them. The parser expects them to only appear
at the top of the file (or the function), and throws an error if they
do.

In the Matlab to Natlab translation, %% comments that don't appear at
the top of the file are meant to be converted to regular % comments.
This was implemented by Andrew Casey in c6679b7ff, back in 2009. But
it looks like there are two "comment" tokens in the translator's
Matlab lexer, and only one of them had this behavior implemented in
its action. This commit just copies copies the action from the first
one to the second.

It seems to fix the issue, but I have no idea if this is the
appropriate way to go about it, or what the implications could be.
I'll revert this commit if something goes wrong.

Migrated from Sable/mclab@fcdd59b48977a3974a46c231366605e2c823b07a