Fix pretty printer bug.
The first argument to String.format is a format string. If we pass in
a literal string with no arguments, and that string happens to contain
a %s (which e.g. a string literal expression could), it'll give an
error. So have wrap() just use the string directly if no arguments
are given.

Migrated from Sable/mclab@96b2824f5c86f0d848f8e9004aede51c2708679d