Removed generics from Shape.
Was not being used. The idea behind having generics which took a class implementing Value as a type parameter was to allow the shape to be reprensented using a BasicMatrixValue or something similar. But that was not being done. It is simpler and possibly less buggy to just use a simple Symbol as is being done right now. A single method growByIndices used the type parameter since the parameter passed to it of type Args also took a type parameter. So method generics was used in that case.